requirement.txt
langchain
openai
streamlit
newsapi-python
python-dotenv
